refactor(#3038): split decision_route.rs into directory modules (verbatim moves)
decision_route.rs becomes the module root (orchestrator +
DecisionResponse/DecisionInput, 139 prod LoC). The 4.4k-line body moves
verbatim into decision_route/:

  repo_card.rs        748 prod  card/review-state PG reads + scope-close tx
  repo_dispatch.rs    650 prod  dispatch lookup/consume/claim CAS repo
  worktree_stale.rs   553 prod  worktree/mainline + stale-review cleanup
  adapters.rs         227 prod  service/event adapters + or_response wrappers
  pending.rs          636 prod  phase 1-2 validation + pending resolution
  accept.rs           703 prod  phase 3a accept branch
  dispute.rs          746 prod  phase 3b dispute branches
  dismiss_finalize.rs 182 prod  phase 3c-4 dismiss + shared finalize

Move fidelity verified by normalized line-multiset diff: the only body
changes are visibility keywords (fn -> pub(super), pub(crate) ->
pub(super); zero external users confirmed by grep) and one module-depth
path adjustment (super::super::kanban -> crate::server::routes::kanban).
Characterization tests travel with their functions (repo_card,
worktree_stale). No route file carries SQL + json! + crate::services
together; every file is <1000 prod LoC.
